admin_api/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose
python-dotenv
pydantic[email]
cryptography
